[#242] Remove unused code and dependencies and restructure project
