--- sample-compose/app/src/main/java/co/nimblehq/sample/compose/ui/base/BaseFragment.kt
@@ -1,90 +1,0 @@
-package co.nimblehq.sample.compose.ui.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.CallSuper
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import androidx.viewbinding.ViewBinding
-import co.nimblehq.common.extensions.hideSoftKeyboard
-import co.nimblehq.sample.compose.ui.common.Toaster
-import co.nimblehq.sample.compose.ui.userReadableMessage
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-abstract class BaseFragment<VB : ViewBinding> : Fragment(), BaseFragmentCallbacks {
-
-    @Inject
-    lateinit var toaster: Toaster
-
-    abstract val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> VB
-
-    private var _binding: ViewBinding? = null
-
-    @Suppress("UNCHECKED_CAST")
-    val binding: VB
-        get() = _binding as VB
-
-    @CallSuper
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        (this as? BaseFragmentCallbacks)?.let { initViewModel() }
-    }
-
-    override fun initViewModel() {}
-
-    override fun setupView() {}
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return bindingInflater.invoke(inflater, container, false).apply {
-            _binding = this
-        }.root
-    }
-
-    @CallSuper
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        (this as? BaseFragmentCallbacks)?.let {
-            setupView()
-            bindViewEvents()
-            bindViewModel()
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-    @CallSuper
-    override fun bindViewEvents() {
-        requireNotNull(view).setOnClickListener {
-            requireActivity().hideSoftKeyboard()
-        }
-    }
-
-    open fun displayError(error: Throwable) {
-        val message = error.userReadableMessage(requireContext())
-        toaster.display(message)
-    }
-
-    protected inline infix fun <T> Flow<T>.bindTo(crossinline action: (T) -> Unit) {
-        with(viewLifecycleOwner) {
-            lifecycleScope.launch {
-                repeatOnLifecycle(Lifecycle.State.STARTED) {
-                    collect { action(it) }
-                }
-            }
-        }
-    }
-}
