+++ sample-compose/app/src/main/java/co/nimblehq/sample/compose/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.sample.compose.ui.screens.compose.theme
+package co.nimblehq.sample.compose.ui.theme
 
 import androidx.compose.ui.graphics.Color
