+++ sample-compose/app/src/main/java/co/nimblehq/sample/compose/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.sample.compose.ui.screens.compose.theme
+package co.nimblehq.sample.compose.ui.theme
 
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.lightColors
