+++ sample-compose/app/src/main/res/navigation/nav_graph_main.xml
@@ -1,17 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_nav_graph"
-    app:startDestination="@id/homeFragment">
-
-    <fragment
-        android:id="@+id/homeFragment"
-        android:name="co.nimblehq.sample.compose.ui.screens.xml.HomeFragment"
-        tools:layout="@layout/fragment_home" />
+    app:startDestination="@id/homeComposeFragment">
 
     <fragment
         android:id="@+id/homeComposeFragment"
-        android:name="co.nimblehq.sample.compose.ui.screens.compose.HomeComposeFragment" />
+        android:name="co.nimblehq.sample.compose.ui.screens.home.HomeComposeFragment" />
 
 </navigation>
