--- sample-compose/app/src/test/java/co/nimblehq/sample/compose/ui/screens/xml/HomeFragmentTest.kt
@@ -1,45 +1,0 @@
-package co.nimblehq.sample.compose.ui.screens.xml
-
-import co.nimblehq.sample.compose.R
-import co.nimblehq.sample.compose.databinding.FragmentHomeBinding
-import co.nimblehq.sample.compose.test.TestNavigatorModule.mockMainNavigator
-import co.nimblehq.sample.compose.test.getPrivateProperty
-import co.nimblehq.sample.compose.test.replace
-import co.nimblehq.sample.compose.ui.BaseFragmentTest
-import dagger.hilt.android.testing.*
-import io.kotest.matchers.nulls.shouldNotBeNull
-import io.kotest.matchers.shouldBe
-import io.mockk.mockk
-import org.junit.*
-
-@HiltAndroidTest
-class HomeFragmentTest : BaseFragmentTest<HomeFragment, FragmentHomeBinding>() {
-
-    private val mockViewModel = mockk<HomeViewModel>(relaxed = true)
-
-    @get:Rule
-    var hiltRule = HiltAndroidRule(this)
-
-    @Before
-    fun setup() {
-        hiltRule.inject()
-    }
-
-    @Test
-    fun `When initializing fragment, it displays the title correctly`() {
-        launchFragment()
-        fragment.binding.tvTitle.text.toString() shouldBe fragment.resources.getString(R.string.app_name)
-    }
-
-    private fun launchFragment() {
-        launchFragmentInHiltContainer<HomeFragment>(
-            onInstantiate = {
-                replace(getPrivateProperty("viewModel"), mockViewModel)
-                navigator = mockMainNavigator
-            }
-        ) {
-            fragment = this
-            fragment.navigator.shouldNotBeNull()
-        }
-    }
-}
